Generating oxxo payment file

+++ srcgen/generators/generate.ts
@@ -1,29 +1,27 @@
 import fs from "fs";
 import path from "path";
 
-export const createFile = (type: string, className: string): void => {
-  const typeToUpper = type.charAt(0).toUpperCase() + type.slice(1);
-  const stubPath = path.join(__dirname, `../stubs/${type}.stub`);
-  const outputPath = path.join(
-    __dirname,
-    `../../${type}s/${className}${typeToUpper}.ts`
-  );
-  console.log('writted.');
 
-  let content = fs.readFileSync(stubPath, "utf8");
+export const createFile = (type: string, className: string, outputPath: string): void => {
+  // Verificar si el archivo ya existe
+  if (fs.existsSync(outputPath)) {
+    console.log(`El archivo type ya existe en ${outputPath}`);
+    return; // Salir de la función si el archivo existe
+  }
+
+  let stubPath = path.join(__dirname, `../stubs/${type}.stub`);
+  if (type === 'resource-controller') {
+    type = 'controller';
+    stubPath = path.join(__dirname, `../stubs/resource-controller.stub`);
+
+  }
+
+
+  let content = fs.readFileSync(stubPath, 'utf8');
   content = content.replace(/\[CLASS_NAME\]/g, className);
   content = content.replace(/\[FILENAME\]/g, className.toLowerCase());
 
   fs.writeFileSync(outputPath, content);
-  console.log(`Created ${outputPath}`);
+  console.log(`Created ${type} at ${outputPath}`);
 };
 
-const className = process.argv[2];
-
-if (!className) {
-  console.error("Please provide a class name");
-  process.exit(1);
-}
-/* createFile("model", className);
-createFile("controller", className);
-createFile("route", className); */
